fix: runcmd on boot missing 'via' keyword
