+++ internal/utils/client.go
@@ -3,11 +3,8 @@ package utils
 import (
 	"context"
 	"fmt"
-	"github.com/ionos-cloud/docker-machine-driver/internal/pointer"
 	"github.com/ionos-cloud/docker-machine-driver/pkg/sdk_utils"
-	"golang.org/x/exp/maps" // General availability soon
 	"gopkg.in/yaml.v3"
-	"strconv"
 	"strings"
 	"time"
 
@@ -51,6 +48,8 @@ func (c *Client) UpdateCloudInitFile(cloudInitYAML string, key string, values []
 		return "", err
 	}
 
+	//someExistingCinit := "I2Nsb3VkLWNvbmZpZwp1c2VyczoKICAtIG5hbWU6IGRlbW8KICAgIHNzaC1hdXRob3JpemVkLWtleXM6CiAgICAgIC0gc3NoLXJzYSBBQUFBQjNOemFDMXljMkVBQUFBREFRQUJBQUFCQVFDdjYwV2p4b00zOUxnUERiaVc3bmUzZ3UxOHEwTklWdjBSRTZyRExOYWwxcXVYWjNucUFsQU5wbDVxbWhEUStHUy9zT3R5Z1NHNC85YWlPQTR2WE81NGsxbUhXTDJpcmp1QjlYYlhyMDArNDR2U2QycS92dFhkR1hoZFNNVGY0L1hLMTdmaktTRy85eTN5RDZubWw2cTlYZ1F4eDlWZi9Ja2FLZGxLMGhiQzFkczArOGg4M1BUYjlkRjNMN2hmM0NoL2dodmo1Kyt0V0pGZEZlRytWSTdFRHVLTkE0ekw4QzVGZFlZV0ZBODhZQW1NOG5kakE1cUNqWlhJSWVadlovejlLcHk2REwwUVo4VDNOc3hSS2FwRVUzbnlpSXVFQW1uOGZibm9zV2Nzb3Z3MElTMUh6NkhzallvNGJ1L2dBODJMV3Qzc2RSVUJaLzdac1ZEM0VMaXAgdXNlckBleGFtcGxlLmNvbQogICAgc3VkbzogWydBTEw9KEFMTCkgTk9QQVNTV0Q6QUxMJ10KICAgIGdyb3Vwczogc3VkbwogICAgc2hlbGw6IC9iaW4vYmFzaApydW5jbWQ6CiAgLSBzZWQgLWkgLWUgJy9eUG9ydC9zL14uKiQvUG9ydCA0NDQ0LycgZXRjL3NzaC9zc2hkX2NvbmZpZw=="
+
 	fmt.Printf("client: Got cloundInit:\n%+v\n", cf)
 
 	if val, ok := cf[key]; ok {
@@ -60,186 +59,13 @@ func (c *Client) UpdateCloudInitFile(cloudInitYAML string, key string, values []
 		cf[key] = values
 	}
 
-	fmt.Printf("client: Set cloundInit:\n%+v\n", cf)
-
 	newCf, err := yaml.Marshal(cf)
 	if err != nil {
 		return "", err
 	}
 	cloudInitYAML = "#cloud-config\n" + string(newCf)
-	return cloudInitYAML, nil
-}
 
-func (c *Client) GetNat(datacenterId, natId string) (*sdkgo.NatGateway, error) {
-	nat, _, err := c.NATGatewaysApi.DatacentersNatgatewaysFindByNatGatewayId(c.ctx, datacenterId, natId).Execute()
-	if err != nil {
-		return nil, sdk_utils.ShortenOpenApiErr(err)
-	}
-	return &nat, nil
-}
-
-func (c *Client) CreateNat(datacenterId string, publicIps []string, lansToGateways map[string][]string, subnet string) (*sdkgo.NatGateway, error) {
-	var lans []sdkgo.NatGatewayLanProperties
-	fmt.Printf("CreateNat(publicIps = %+v, lansMap = %+v, subnet = %s)\n", publicIps, lansToGateways, subnet)
-
-	err := c.createLansIfNotExist(datacenterId, maps.Keys(lansToGateways))
-	if err != nil {
-		return nil, err
-	}
-	time.Sleep(10 * time.Second)
-	for lanId, gatewayIps := range lansToGateways {
-		id, err := strconv.ParseInt(lanId, 10, 32)
-		if err != nil {
-			return nil, err
-		}
-		// Unpack the map into NatGatewayLanProperties objects. https://api.ionos.com/docs/cloud/v6/#tag/NAT-Gateways/operation/datacentersNatgatewaysPost
-		var ptrGatewayIps *[]string = nil
-		if len(gatewayIps) > 1 || gatewayIps[0] != "" {
-			// We do this check so that we don't set the GatewayIps property if it's empty. If the property is empty, a gateway IP is generated by the API.
-			ptrGatewayIps = &gatewayIps
-		}
-		lans = append(lans, sdkgo.NatGatewayLanProperties{Id: pointer.To(int32(id)), GatewayIps: ptrGatewayIps})
-		fmt.Printf("Created a NatGatewayLanProperties obj with Id: %d, GatewayIps: %+v\n", id, gatewayIps)
-	}
-
-	// Configure default NAT Rules
-	rules, err := c.MakeNatRules(
-		publicIps[0],
-		subnet,
-		subnet,
-		[]sdkgo.TargetPortRange{
-			//{
-			//	Start: pointer.To(int32(22)), // TODO: Investigate if rancher/rke/issues/212 is a complete list
-			//	End:   pointer.To(int32(22)),
-			//},
-			//{
-			//	Start: pointer.To(int32(6443)),
-			//	End:   pointer.To(int32(6443)),
-			//},
-			//{
-			//	Start: pointer.To(int32(2379)),
-			//	End:   pointer.To(int32(2380)),
-			//},
-			//{
-			//	Start: pointer.To(int32(10250)),
-			//	End:   pointer.To(int32(10252)),
-			//},
-			//{
-			//	Start: pointer.To(int32(10255)), // Investigate 10255 if needed
-			//	End:   pointer.To(int32(10256)),
-			//},
-			//{
-			//	Start: pointer.To(int32(8472)),
-			//	End:   pointer.To(int32(8472)),
-			//},
-			//{
-			//	Start: pointer.To(int32(30000)), // Worker Nodes ports. Maybe worker nodes should have different port config than others?
-			//	End:   pointer.To(int32(32767)),
-			//},
-			{
-				Start: pointer.To(int32(21)),
-				End:   pointer.To(int32(40000)),
-			},
-		},
-	)
-
-	if err != nil {
-		return nil, err
-	}
-
-	nat, resp, err := c.NATGatewaysApi.DatacentersNatgatewaysPost(c.ctx, datacenterId).NatGateway(
-		sdkgo.NatGateway{
-			Properties: &sdkgo.NatGatewayProperties{
-				Name:      pointer.To("NAT Docker Machine"),
-				PublicIps: &publicIps,
-				Lans:      &lans,
-			},
-			Entities: &sdkgo.NatGatewayEntities{
-				Rules:    &sdkgo.NatGatewayRules{Items: rules},
-				Flowlogs: nil,
-			},
-		},
-	).Execute()
-
-	if err != nil {
-		return nil, err
-	}
-
-	fmt.Printf("created nat: %+v\n", nat)
-
-	err = c.waitTillProvisioned(resp.Header.Get("location"))
-
-	return &nat, err
-}
-
-// MakeNatRules returns uncreated NatGatewayRules, which are meant to be sent as "entities" for the NAT Post request
-func (c *Client) MakeNatRules(publicIp, srcSubnet, targetSubnet string, portRanges []sdkgo.TargetPortRange) (*[]sdkgo.NatGatewayRule, error) {
-	properties := &sdkgo.NatGatewayRuleProperties{
-		Name: pointer.To("Docker Machine NAT Rule"),
-		//Type:         pointer.To(sdkgo.NatGatewayRuleType("SNAT")),
-		//Protocol:     pointer.To(sdkgo.NatGatewayRuleProtocol("ALL")),
-		SourceSubnet: &srcSubnet,
-		//TargetSubnet: &targetSubnet,
-		PublicIp: &publicIp,
-	}
-
-	rules := make([]sdkgo.NatGatewayRule, 0)
-	//for _, portRange := range portRanges {
-	//properties.TargetPortRange = &portRange
-	rule := sdkgo.NatGatewayRule{Properties: properties}
-	rules = append(rules, rule)
-	//}
-
-	return &rules, nil
-}
-
-func (c *Client) createLansIfNotExist(datacenterId string, lanIds []string) error {
-	for _, lanid := range lanIds {
-		_, resp, err := c.LANsApi.DatacentersLansFindById(c.ctx, datacenterId, lanid).Execute()
-		if resp.StatusCode == 404 {
-			// Run this before err check, as 404s throws an err.
-			fmt.Printf("Creating LAN %s for NAT\n", lanid)
-			_, err := c.CreateLan(datacenterId, "Docker Machine LAN (NAT)", false)
-			if err != nil {
-				return err
-			}
-			continue // breakpoint
-		}
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (c *Client) RemoveNat(datacenterId, natId string) error {
-	_, err := c.NATGatewaysApi.DatacentersNatgatewaysDelete(c.ctx, datacenterId, natId).Execute()
-	if err != nil {
-		return sdk_utils.ShortenOpenApiErr(err)
-	}
-	return nil
-}
-
-// UpdateCloudInitFile will try to unmarshal the cloud config YAML string, append the given values to the selected key, and then marshal it back into a YAML string which is returned
-func (c *Client) UpdateCloudInitFile(cloudInitYAML string, key string, values []interface{}) (string, error) {
-	var cf map[string]interface{}
-	cf = make(map[string]interface{})
-	if err := yaml.Unmarshal([]byte(cloudInitYAML), &cf); err != nil {
-		return "", err
-	}
-
-	if val, ok := cf[key]; ok {
-		u := val.([]interface{})
-		cf[key] = append(u, values...)
-	} else {
-		cf[key] = values
-	}
-
-	newCf, err := yaml.Marshal(cf)
-	if err != nil {
-		return "", err
-	}
-	cloudInitYAML = "#cloud-config\n" + string(newCf)
+	fmt.Printf("client: Set cloundInit:\n\n%+v\n", cloudInitYAML)
 
 	return cloudInitYAML, nil
 }
